Load logfile.conf first to help debugging
Conf files in $COLLECTD_CONF_DIR are loaded in alphabetical order.

If there is an error loading the python plugin, the logfile plugin is not
loaded.
To help debugging, the logfile is should be loaded first.

Changing logfile.conf to 01-logfile.conf should fix this issue by making
collectd load the logfile.conf first

Closes-Bug: #1657859

Change-Id: I1f748d61d6d9a291bcc32476ac17bb0d60fdc573