Move tests from aodh/test_plugin to aodh/test_sender
A number of tests for aodh.sender.Sender are in
aodh/test_plugin.py. Since aodh/test_sender.py was recently
added, it makes sense to move the sender tests to this new
location.

The following tests were moved:
* test_get_alarm_state_severity_low
* test_get_alarm_state_severity_moderate
* test_get_alarm_state_severity_critical

The tests were refactored to use self.sender that is defined
in TestSender/setUp().

Change-Id: I337c37b14f593183a46b8fd4b077c5b47bc791bc