gnocchi: Remove archive-policy when creating metrics
Archive policy rules are created by the operator, and assign
archive policies to metrics at creation based on certain rules,
if there isn't an archive policy explicitly assigned to the
metric already.

This patch removes the explicit archive policy parameter when
creating the metrics, this means that the archive policy rules
defined by the oparator are respected.

Change-Id: I562f62e1236c7c01c79a5e465ebec103c7bb07bd