Fix re-authorisation problem for AODH plugin
Previously AODH plugin in '_create_request_url' function
was not only creating request URL but also it was performing
put-request. It caused a problem because this particular request
was not protected for authorization-expired scenario.

Now actual url generation and perform-request parts are
separated like for gnocchi plugin and authorization is
renewed in problematic scenarios.

Change-Id: Idf72b1ff381fd97acfeb0c067fc86c44f59da2fe
Signed-off-by: Radoslaw Jablonski <radoslawx.jablonski@intel.com>